requirements/base.txt
Django==3.1.7
Pillow==8.1.2
psycopg2==2.8.5
unidecode==1.2
django-ckeditor==6.0.0
